waktu

+++ database/migrations/2026_02_15_030743_drop_mata_pelajaran_from_gurus_table.php
@@ -4,19 +4,21 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+return new class extends Migration
+{
     public function up(): void
     {
-        Schema::table('gurus', function (Blueprint $table) {
-            $table->dropColumn('mata_pelajaran');
-        });
+        if (Schema::hasColumn('gurus', 'mata_pelajaran')) {
+
+            Schema::table('gurus', function (Blueprint $table) {
+                $table->dropColumn('mata_pelajaran');
+            });
+
+        }
     }
 
     public function down(): void
     {
-        Schema::table('gurus', function (Blueprint $table) {
-            $table->string('mata_pelajaran');
-        });
+        //
     }
 };
-
